# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/milestones/{id}": {
      "get": {
        "summary": "Get milestone by ID",
        "tags": [
          "Milestones"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Milestone retrieved successfully"
          }
        }
      },
      "put": {
        "summary": "Update milestone",
        "tags": [
          "Milestones"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Milestone updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete milestone",
        "tags": [
          "Milestones"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Milestone deleted successfully"
          }
        }
      }
    },
    "/api/milestones": {
      "get": {
        "summary": "Get all milestones",
        "tags": [
          "Milestones"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Milestones retrieved successfully"
          }
        }
      },
      "post": {
        "summary": "Create new milestone",
        "tags": [
          "Milestones"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Milestone created successfully"
          }
        }
      }
    },
    "/api/tasks/{id}": {
      "get": {
        "summary": "Get task by ID",
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Task retrieved successfully"
          }
        }
      },
      "put": {
        "summary": "Update task",
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Task updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete task",
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted successfully"
          }
        }
      }
    },
    "/api/tasks": {
      "get": {
        "summary": "Get all tasks",
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Tasks retrieved successfully"
          }
        }
      },
      "post": {
        "summary": "Create new task",
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Task created successfully"
          }
        }
      }
    },
...
  },
...
}